--- src/main/java/finalmission/dto/request/MovieCreateRequest.java
@@ -1,5 +1,0 @@
-package finalmission.dto.request;
-
-public record MovieCreateRequest(String name, String description) {
-
-}
