--- src/main/java/finalmission/dto/request/MovieReservationCreateRequest.java
@@ -1,4 +1,0 @@
-package finalmission.dto.request;
-
-public record MovieReservationCreateRequest(String memberName, Long movieSlotId, Integer seat) {
-}
