--- src/main/java/finalmission/dto/request/MovieSlotCreateRequest.java
@@ -1,8 +1,0 @@
-package finalmission.dto.request;
-
-import java.time.LocalDate;
-import java.time.LocalTime;
-
-public record MovieSlotCreateRequest(Long movieId, LocalDate date, LocalTime startAt, Integer seats) {
-
-}
