--- src/main/java/finalmission/dto/response/MovieCreateResponse.java
@@ -1,5 +1,0 @@
-package finalmission.dto.response;
-
-public record MovieCreateResponse(Long id, String name, String description) {
-
-}
