--- src/main/java/finalmission/dto/response/MovieReservationCreateResponse.java
@@ -1,4 +1,0 @@
-package finalmission.dto.response;
-
-public record MovieReservationCreateResponse(Long id, String memberName, String movieName, Integer seats) {
-}
