--- src/main/java/finalmission/dto/response/MovieReservationReadResponse.java
@@ -1,12 +1,0 @@
-package finalmission.dto.response;
-
-import java.time.LocalDate;
-import java.time.LocalTime;
-
-public record MovieReservationReadResponse(Long id,
-                                           String movieName,
-                                           LocalDate date,
-                                           LocalTime startAt,
-                                           Integer seat
-) {
-}
