--- src/main/java/finalmission/dto/response/MovieSlotCreateResponse.java
@@ -1,7 +1,0 @@
-package finalmission.dto.response;
-
-import java.time.LocalDate;
-import java.time.LocalTime;
-
-public record MovieSlotCreateResponse(Long id, String movieName, LocalDate date, LocalTime startAt, Integer seats) {
-}
