--- src/main/java/finalmission/dto/response/MovieSlotReadResponse.java
@@ -1,12 +1,0 @@
-package finalmission.dto.response;
-
-import java.time.LocalDate;
-import java.time.LocalTime;
-
-public record MovieSlotReadResponse(Long id,
-                                    String movieName,
-                                    LocalDate date,
-                                    LocalTime startAt,
-                                    Integer totalSeats,
-                                    Integer leftSeats) {
-}
