+++ src/main/java/finalmission/movie/dto/request/MovieCreateRequest.java
@@ -1,0 +1,5 @@
+package finalmission.movie.dto.request;
+
+public record MovieCreateRequest(String name, String description) {
+
+}
