+++ src/main/java/finalmission/movie/dto/request/MovieReservationCreateRequest.java
@@ -1,0 +1,4 @@
+package finalmission.movie.dto.request;
+
+public record MovieReservationCreateRequest(String memberName, Long movieSlotId, Integer seat) {
+}
