+++ src/main/java/finalmission/movie/dto/request/MovieSlotCreateRequest.java
@@ -1,0 +1,8 @@
+package finalmission.movie.dto.request;
+
+import java.time.LocalDate;
+import java.time.LocalTime;
+
+public record MovieSlotCreateRequest(Long movieId, LocalDate date, LocalTime startAt, Integer seats) {
+
+}
