+++ src/main/java/finalmission/movie/dto/response/MovieCreateResponse.java
@@ -1,0 +1,5 @@
+package finalmission.movie.dto.response;
+
+public record MovieCreateResponse(Long id, String name, String description) {
+
+}
