+++ src/main/java/finalmission/movie/dto/response/MovieReservationCreateResponse.java
@@ -1,0 +1,4 @@
+package finalmission.movie.dto.response;
+
+public record MovieReservationCreateResponse(Long id, String memberName, String movieName, Integer seats) {
+}
