+++ src/main/java/finalmission/movie/dto/response/MovieReservationReadResponse.java
@@ -1,0 +1,12 @@
+package finalmission.movie.dto.response;
+
+import java.time.LocalDate;
+import java.time.LocalTime;
+
+public record MovieReservationReadResponse(Long id,
+                                           String movieName,
+                                           LocalDate date,
+                                           LocalTime startAt,
+                                           Integer seat
+) {
+}
