+++ src/main/java/finalmission/movie/dto/response/MovieSlotReadResponse.java
@@ -1,0 +1,12 @@
+package finalmission.movie.dto.response;
+
+import java.time.LocalDate;
+import java.time.LocalTime;
+
+public record MovieSlotReadResponse(Long id,
+                                    String movieName,
+                                    LocalDate date,
+                                    LocalTime startAt,
+                                    Integer totalSeats,
+                                    Integer leftSeats) {
+}
