+++ src/main/java/finalmission/movie/repository/MovieRepository.java
@@ -1,0 +1,7 @@
+package finalmission.movie.repository;
+
+import finalmission.movie.entity.Movie;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface MovieRepository extends JpaRepository<Movie, Long> {
+}
