--- src/main/java/finalmission/repository/MovieRepository.java
@@ -1,7 +1,0 @@
-package finalmission.repository;
-
-import finalmission.entity.Movie;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface MovieRepository extends JpaRepository<Movie, Long> {
-}
